fix(): Fixed the type of score(dc)

+++ server/src/main/kotlin/delta/codecharacter/server/user/public_user/PublicUserEntity.kt
@@ -18,6 +18,6 @@ data class PublicUserEntity(
     val losses: Int,
     val ties: Int,
     val isActivated: Boolean = true,
-    val score: Int,
+    val score: Double,
     val challengesCompleted: Array<String>?
 )
